+++ src/App.tsx
@@ -1,16 +1,16 @@
 import React, { Component } from 'react';
-import Home from './Home';
-import Login from './Login';
-import EditProfile from './EditProfile';
-import ForgotPassword from './ForgotPassword';
-import ResetPassword from './ResetPassword';
-import ChangePassword from './ChangePassword';
-import VerifyAccount from './VerifyAccount';
+import Home from './routes/Home';
+import Login from './routes/Login';
+import EditProfile from './routes/EditProfile';
+import ForgotPassword from './routes/ForgotPassword';
+import ResetPassword from './routes/ResetPassword';
+import ChangePassword from './routes/ChangePassword';
+import VerifyAccount from './routes/VerifyAccount';
 import Admin from './routes/admin';
-import Privacy from './Privacy';
-import Terms from './Terms';
-import Docs from './Docs';
-import Faq from './FAQ';
+import Privacy from './routes/Privacy';
+import Terms from './routes/Terms';
+import Docs from './routes/Docs';
+import Faq from './routes/FAQ';
 import BeepAppBar from './AppBar';
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import { UserContext } from './UserContext';
