+++ src/routes/FAQ.tsx
