+++ src/routes/admin/reports/Report.tsx
@@ -1,7 +1,5 @@
 import React, { useEffect } from 'react'
-import { useSelector, useDispatch } from 'react-redux';
 import { useParams } from "react-router-dom";
-import { userSelector, fetchUser } from '../../../store/slices/users';
 
 import { Heading3, Body1 } from '../../../components/Typography';
 
@@ -9,8 +7,6 @@ function Report(props) {
 
     const { reportId } = useParams<{reportId: string}>();
 
-    // const dispatch = useDispatch();
-    // const user = useSelector(userSelector(userId));
 
     useEffect(() => {
         // dispatch(fetchUser(userId));
